EchoEyes/requirements.txt
python-telegram-bot==21.6
requests>=2.32.3
web3==6.19.0
solana==0.36.9
solders==0.20.0
APScheduler==3.10.4
watchdog==4.0.2
httpx>=0.27.2
typing-extensions>=4.12.2
